# perf(export): do not allocate arrays if resource has no pending async attributes