--- a/packages/opentelemetry-sdk-trace-base/src/export/BatchSpanProcessorBase.ts
+++ b/packages/opentelemetry-sdk-trace-base/src/export/BatchSpanProcessorBase.ts
@@ -195,19 +195,24 @@ export abstract class BatchSpanProcessorBase<T extends BufferConfig>
               );
             }
           });
-        const pendingResources = spans
-          .map(span => span.resource)
-          .filter(resource => resource.asyncAttributesPending);
+
+        let pendingResources: Array<Promise<void>> | null = null;
+        for (let i = 0, len = spans.length; i < len; i++) {
+          const span = spans[i];
+          if (
+            span.resource.asyncAttributesPending &&
+            span.resource.waitForAsyncAttributes
+          ) {
+            pendingResources ??= [];
+            pendingResources.push(span.resource.waitForAsyncAttributes());
+          }
+        }
 
         // Avoid scheduling a promise to make the behavior more predictable and easier to test
-        if (pendingResources.length === 0) {
+        if (pendingResources === null) {
           doExport();
         } else {
-          Promise.all(
-            pendingResources.map(
-              resource => resource.waitForAsyncAttributes?.()
-            )
-          ).then(doExport, err => {
+          Promise.all(pendingResources).then(doExport, err => {
             globalErrorHandler(err);
             reject(err);
           });