## Which problem is this PR solving?

OpenTelemetry is producing too much overhead for scenarios with significant load. One of the example is BatchSpanExporter, that is allocating array of batch size only to check is it has any pending attributes. 

Fixes # (issue)

## Short description of the changes

Replaced map & filter with a simple loop.

## Type of change

Please delete options that are not relevant.

- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

## How Has This Been Tested?

Benchmarks:

Before 
create spans (10 attributes) x 689,041 ops/sec ±1.54% (92 runs sampled)
BatchSpanProcessor process span x 660,065 ops/sec ±1.08% (95 runs sampled)

After:
create spans (10 attributes) x 688,731 ops/sec ±1.15% (94 runs sampled)
BatchSpanProcessor process span x 673,550 ops/sec ±1.01% (93 runs sampled)

- [ ] Test A

## Checklist:

- [ ] Followed the style guidelines of this project
- [ ] Unit tests have been added
- [ ] Documentation has been updated