@Samuron 2024-03-17
I am not sure that the benchmarks are very representational though, they fluctuate a lot and the numbers are suspiciously close to just creating spans